pyproject.toml
[project]
name = "check-my-homework"
version = "0.1.0"
description = "A simple Python script that uses LLMs to check homework PDFs for mistakes"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Karan Bania" },
    { name = "Claude Opus 4.5" },
]
dependencies = [
    "google-genai>=1.0.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
]

[project.scripts]
check-homework = "check:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
